requirements.txt
customtkinter
sqlalchemy
dnspython
matplotlib
apscheduler
plyer
reportlab
requests